+++ src/diagnostics/build.ts
@@ -15,7 +15,7 @@ import * as gnu_ld from '@cmt/diagnostics/gnu-ld';
 import * as mvsc from '@cmt/diagnostics/msvc';
 import * as iar from '@cmt/diagnostics/iar';
 import * as iwyu from '@cmt/diagnostics/iwyu';
-import { FileDiagnostic, RawDiagnostic, RawDiagnosticParser } from '@cmt/diagnostics/util';
+import { FileDiagnostic, RawDiagnostic, RawDiagnosticParser, diagnosticSeverity } from '@cmt/diagnostics/util';
 import { ConfigurationReader } from '@cmt/config';
 import { fs } from '@cmt/pr';
 
@@ -36,6 +36,19 @@ export class CompileOutputConsumer implements OutputConsumer {
 
     compilers = new Compilers();
 
+    private readonly _onDiagnosticEmitter = new vscode.EventEmitter<RawDiagnostic>();
+
+    /**
+     * Event fired when a new diagnostic is parsed from compiler output
+     */
+    get onDiagnostic() {
+        return this._onDiagnosticEmitter.event;
+    }
+
+    dispose() {
+        this._onDiagnosticEmitter.dispose();
+    }
+
     // Defer all output to the `error` method
     output(line: string) {
         this.error(line);
@@ -43,7 +56,12 @@ export class CompileOutputConsumer implements OutputConsumer {
 
     error(line: string) {
         for (const cand in this.compilers) {
-            if (this.compilers[cand].handleLine(line)) {
+            const parser = this.compilers[cand];
+            const countBefore = parser.diagnostics.length;
+            if (parser.handleLine(line)) {
+                if (parser.diagnostics.length > countBefore) {
+                    this._onDiagnosticEmitter.fire(parser.diagnostics[parser.diagnostics.length - 1]);
+                }
                 break;
             }
         }
@@ -63,24 +81,6 @@ export class CompileOutputConsumer implements OutputConsumer {
         const diags_by_file = new Map<string, vscode.Diagnostic[]>();
         const linkerHandler = this.createLinkerDiagnosticsHandler(basePaths);
 
-        const severity_of = (p: string) => {
-            switch (p) {
-                case 'warning':
-                    return vscode.DiagnosticSeverity.Warning;
-                case 'catastrophic error':
-                case 'fatal error':
-                case 'error':
-                    return vscode.DiagnosticSeverity.Error;
-                case 'note':
-                case 'info':
-                case 'remark':
-                    return vscode.DiagnosticSeverity.Information;
-            }
-            // tslint:disable-next-line
-            console.warn('Unknown diagnostic severity level: ' + p);
-            return undefined;
-        };
-
         const by_source = {
             GCC: this.compilers.gcc.diagnostics,
             MSVC: this.compilers.msvc.diagnostics,
@@ -99,7 +99,7 @@ export class CompileOutputConsumer implements OutputConsumer {
             for (const raw_diag of diags) {
                 await linkerHandler.collect(raw_diag, source, arrs.length);
                 const filepath = await this.resolvePath(raw_diag.file, basePaths);
-                const severity = severity_of(raw_diag.severity);
+                const severity = diagnosticSeverity(raw_diag.severity);
                 if (severity === undefined) {
                     continue;
                 }
@@ -276,8 +276,16 @@ export class CMakeBuildConsumer extends proc.CommandConsumer implements vscode.D
 
     readonly compileConsumer: CompileOutputConsumer;
 
+    /**
+     * Event fired when a new diagnostic is parsed from compiler output
+     */
+    get onDiagnostic() {
+        return this.compileConsumer.onDiagnostic;
+    }
+
     dispose() {
         this._onProgressEmitter.dispose();
+        this.compileConsumer.dispose();
     }
 
     error(line: string) {